refactor: 리뷰 요약 컴포넌트 위치 정리
