--- apps/web/src/widgets/place-review-summary/index.ts
@@ -1,7 +1,0 @@
-export { mockPlaceReviewSummary } from './model/mock-data';
-export type {
-  PlaceReviewMetric,
-  PlaceReviewMetricType,
-  PlaceReviewSummary,
-  PlaceReviewVariant,
-} from './model/types';
